+++ src/icons/food-beverage/index.ts
@@ -1,3 +1,17 @@
+import BeerMug from '@/src/icons/food-beverage/BeerMug';
+import Cake from '@/src/icons/food-beverage/Cake';
+import Cutlery from '@/src/icons/food-beverage/Cutlery';
 import EatenApple from '@/src/icons/food-beverage/EatenApple';
+import FoodTiffin from '@/src/icons/food-beverage/FoodTiffin';
+import GreenTea from '@/src/icons/food-beverage/GreenTea';
+import PotatoChips from '@/src/icons/food-beverage/PotatoChips';
 
-export { EatenApple };
+export {
+    BeerMug,
+    Cake,
+    Cutlery,
+    EatenApple,
+    FoodTiffin,
+    GreenTea,
+    PotatoChips
+};