+++ src/icons/household/index.ts
@@ -1,3 +1,4 @@
+import Appliances from '@/src/icons/household/Appliances';
 import Sofa from '@/src/icons/household/Sofa';
 
-export { Sofa };
+export { Appliances, Sofa };