+++ src/icons/sports/index.ts
@@ -1,3 +1,4 @@
+import Basketball from '@/src/icons/sports/Basketball';
 import Dart from '@/src/icons/sports/Dart';
 
-export { Dart };
+export { Basketball, Dart };